Typo Check Feature #9
In this commit, I added a typo check feature for text area.
When user typed in wrong instruction, an alert will pop up immediately and tell them it is wrong and they need to fix it.
The feature is not well-developed. I put the rest implementation in comment for MIPS team to implement further.